add generate error vector, generate polynomial
add an assert to specify that eta*poly_size/4 needs to be equal to the `prf` output length, which is hardcoded at 192 for `prf_3` (and 128 for `prf_2`. this may be more relevant later, but we're going to use two asserts, before and after the call to `cbd` because in cbd it's generic, whereas this tells you exactly what the parameters need to be.