set `n=256`
note that the polynomial size needs to be 256 if `eta_1=3` since `3*256/4 = 192`, and this is the `prf` output size.

further, it is currently not possible to set both `n=256` and `q=3329`, despite having an `n`^th root of unity in such a case. we don't have a 2n^th root of unity, and out current implementation of the `ntt` requires a `2n`th root of unity.