use arbitrary precision to fix encode
the variable `t` needs to be an arbitrary precision integer to allow for the full bytes to be read into it. using a `u128` is far too small.

we also allow an option just to mod the coefficients modulo `q` so they're in [0,q-1]. this is not strictly necessary, but may be useful in the future.