use correct `prf` type for `generate_polynomial`
make sure that whenever we use `prf`, we use the right type depending on `eta`. probably we could just pass a parameter to explicitly set the `prf` output size, but this should be fine for now and is more precise, because there are really only two sizes for `eta=2` and `eta=3`, and we're just not allowing anything else.